% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion_menu.R
\name{accordion_menu}
\alias{accordion_menu}
\title{accordion_menu
This function is a Ui wrapper for a sidebar menu packed as accordion. This allows to
increase the complexity of the app by hiding most of the content by default.}
\usage{
accordion_menu(title, ..., active = F, disabled = F)
}
\description{
accordion_menu
This function is a Ui wrapper for a sidebar menu packed as accordion. This allows to
increase the complexity of the app by hiding most of the content by default.
}
